fix(cloudsmith): add nil checks after API calls and move test helpers
