+++ pkg/integrations/cloudsmith/common.go
@@ -10,7 +10,6 @@ import (
 	"github.com/mitchellh/mapstructure"
 	"github.com/superplanehq/superplane/pkg/configuration"
 	"github.com/superplanehq/superplane/pkg/core"
-	"github.com/superplanehq/superplane/test/support/contexts"
 )
 
 const (
@@ -124,30 +123,3 @@ func defaultHandleWebhook(ctx core.WebhookRequestContext) (int, *core.WebhookRes
 	return http.StatusOK, nil, nil
 }
 
-func cloudsmithPackageExecutionContext(httpContext *contexts.HTTPContext, executionState *contexts.ExecutionStateContext) core.ExecutionContext {
-	return cloudsmithPackageExecutionContextWithConfiguration(
-		httpContext,
-		executionState,
-		map[string]any{
-			"repository": "acme/production",
-			"package":    "pkg_123",
-		},
-	)
-}
-
-func cloudsmithPackageExecutionContextWithConfiguration(
-	httpContext *contexts.HTTPContext,
-	executionState *contexts.ExecutionStateContext,
-	configuration map[string]any,
-) core.ExecutionContext {
-	return core.ExecutionContext{
-		Configuration: configuration,
-		HTTP:          httpContext,
-		Integration: &contexts.IntegrationContext{
-			Configuration: map[string]any{
-				"apiKey": "test-key",
-			},
-		},
-		ExecutionState: executionState,
-	}
-}
