+++ pkg/integrations/cloudsmith/example_output_tag_package.json
@@ -1,33 +1,37 @@
 {
-  "cdn_url": "https://dl.cloudsmith.io/basic/example-owner/example-repo/docker/example-package-id.manifest.json",
-  "checksum_md5": "00000000000000000000000000000000",
-  "checksum_sha1": "0000000000000000000000000000000000000000",
-  "checksum_sha256": "0000000000000000000000000000000000000000000000000000000000000000",
-  "checksum_sha512": "0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-  "description": "Example tagged package payload",
-  "display_name": "example-app",
-  "filename": "example-app",
-  "format": "docker",
-  "is_quarantined": false,
-  "is_sync_completed": true,
-  "name": "example-app",
-  "namespace": "example-owner",
-  "repository": "example-repo",
-  "self_html_url": "https://cloudsmith.io/~example-owner/repos/example-repo/packages/detail/docker/example-app/example-version/",
-  "self_url": "https://api.cloudsmith.io/v1/packages/example-owner/example-repo/example-package-id/",
-  "self_webapp_url": "https://app.cloudsmith.com/example-owner/r/example-repo/docker/example-app/example-version/example-package-id",
-  "size": 123456,
-  "size_str": "",
-  "slug": "example-app-abc1",
-  "slug_perm": "example-package-id",
-  "status": 4,
-  "status_str": "Completed",
-  "summary": "",
-  "tags": {
-    "info": ["example"],
-    "version": ["latest"]
+  "data": {
+    "cdn_url": "https://dl.cloudsmith.io/basic/example-owner/example-repo/docker/example-package-id.manifest.json",
+    "checksum_md5": "00000000000000000000000000000000",
+    "checksum_sha1": "0000000000000000000000000000000000000000",
+    "checksum_sha256": "0000000000000000000000000000000000000000000000000000000000000000",
+    "checksum_sha512": "0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
+    "description": "Example tagged package payload",
+    "display_name": "example-app",
+    "filename": "example-app",
+    "format": "docker",
+    "is_quarantined": false,
+    "is_sync_completed": true,
+    "name": "example-app",
+    "namespace": "example-owner",
+    "repository": "example-repo",
+    "self_html_url": "https://cloudsmith.io/~example-owner/repos/example-repo/packages/detail/docker/example-app/example-version/",
+    "self_url": "https://api.cloudsmith.io/v1/packages/example-owner/example-repo/example-package-id/",
+    "self_webapp_url": "https://app.cloudsmith.com/example-owner/r/example-repo/docker/example-app/example-version/example-package-id",
+    "size": 123456,
+    "size_str": "",
+    "slug": "example-app-abc1",
+    "slug_perm": "example-package-id",
+    "status": 4,
+    "status_str": "Completed",
+    "summary": "",
+    "tags": {
+      "info": ["example"],
+      "version": ["latest"]
+    },
+    "tags_immutable": {},
+    "uploaded_at": "2026-01-01T00:00:00Z",
+    "version": "example-version"
   },
-  "tags_immutable": {},
-  "uploaded_at": "2026-01-01T00:00:00Z",
-  "version": "example-version"
+  "timestamp": "2026-06-19T10:21:49.411256712Z",
+  "type": "cloudsmith.package.tagged"
 }
